Move formatters back in header files
