+++ cpp/arcticdb/entity/ref_key.cpp
@@ -5,31 +5,7 @@
  * As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
  */
 
-#include <fmt/format.h>
 #include <arcticdb/entity/ref_key.hpp>
-
-template<>
-struct fmt::formatter<arcticdb::entity::RefKey>
-{
-    template<typename ParseContext>
-    constexpr auto parse(ParseContext& ctx) { return ctx.begin(); }
-
-    template<typename FormatContext>
-    auto format(const arcticdb::entity::RefKey& number, FormatContext& ctx) {
-        return fmt::format_to(ctx.out(), "{}:{}", number.type(), number.id());
-    }
-};
-
-//TODO this is operating on the pretty-printed version and is needlessly inefficient
-namespace std {
-    template<>
-    struct hash<arcticdb::entity::RefKey> {
-        inline arcticdb::HashedValue operator()(const arcticdb::entity::RefKey &k) const noexcept {
-            auto view = k.view();
-            return arcticdb::hash(const_cast<uint8_t * >(reinterpret_cast<const uint8_t *>(view.data())), view.size());
-        }
-    };
-}
 
 namespace arcticdb::entity {
